+++ packages/angular-query-persist-client/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tanstack/angular-query-persist-client
 
+## 5.100.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.100.12
+  - @tanstack/query-persist-client-core@5.100.12
+
+## 5.100.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.100.11
+  - @tanstack/query-persist-client-core@5.100.11
+
+## 5.100.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/angular-query-experimental@5.100.10
+  - @tanstack/query-persist-client-core@5.100.10
+
 ## 5.100.9
 
 ### Patch Changes