+++ packages/preact-query-devtools/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tanstack/preact-query-devtools
 
+## 5.100.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.100.12
+  - @tanstack/query-devtools@5.100.12
+
+## 5.100.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/preact-query@5.100.11
+  - @tanstack/query-devtools@5.100.11
+
+## 5.100.10
+
+### Patch Changes
+
+- Updated dependencies [[`4d130b9`](https://github.com/TanStack/query/commit/4d130b992d2ac396d75f921bfc530dd3a9d50577)]:
+  - @tanstack/query-devtools@5.100.10
+  - @tanstack/preact-query@5.100.10
+
 ## 5.100.9
 
 ### Patch Changes