+++ packages/query-async-storage-persister/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @tanstack/query-async-storage-persister
 
+## 5.100.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.100.12
+  - @tanstack/query-persist-client-core@5.100.12
+
+## 5.100.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.100.11
+  - @tanstack/query-persist-client-core@5.100.11
+
+## 5.100.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-core@5.100.10
+  - @tanstack/query-persist-client-core@5.100.10
+
 ## 5.100.9
 
 ### Patch Changes